+++ CHANGELOG.md
@@ -19,4 +19,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Fix entry point caching error when trying to get the value of an unset field
 * Fix incorrect uppercase check when calculating fuzzy search scores
 * Fix NodeGraph editor losing reference to graph when entering play mode
+* Fix input sockets still being editable when they are linked to